// web/lib/types/TenantAccessSummary.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AuditWebhookSummary } from "./AuditWebhookSummary";
import type { RolePermissionSummary } from "./RolePermissionSummary";
import type { TenantInfo } from "./TenantInfo";
import type { TenantInvitationSummary } from "./TenantInvitationSummary";
import type { TenantMemberSummary } from "./TenantMemberSummary";

export type TenantAccessSummary = { tenant: TenantInfo, members: Array<TenantMemberSummary>, invitations: Array<TenantInvitationSummary>, role_matrix: Array<RolePermissionSummary>, audit_webhooks: Array<AuditWebhookSummary>, };