// web/lib/types/VolumeSummary.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type VolumeSummary = { id: string, name: string, host_id: string, host_name: string, size_mb: number, backend: string, host_path: string | null, status: string, created_at: string, };